+++ section/footer.php
@@ -40,3 +40,39 @@ THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR I
           
           </section>
         </footer><!-- /#footer -->
+        
+        
+	</div><!-- /#master-wrap-->
+	
+	
+	
+	
+	
+	    <!-- Modal Popup Contact Form - Replace this with your own form -->
+	    <?php include("section/contact-form.php"); ?>
+
+
+			
+	
+	
+		<!-- Scripts -->
+        <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.min.js"></script>
+        <script>window.jQuery || document.write('<script src="js/jquery-1.9.1.min.js"><\/script>')</script>
+        <script src="js/script.js"></script>
+        <script src="js/custom.js"></script>
+        <script src="js/smooth-scroll.min.js"></script><!-- Smooth Scroll -->
+        <script src="js/wufoo.js"></script><!-- Form control -->
+        
+
+		<!-- Analytics -->
+        <script>
+            var _gaq=[['_setAccount','UA-XXXXX-X'],['_trackPageview']];
+            (function(d,t){var g=d.createElement(t),s=d.getElementsByTagName(t)[0];
+            g.src=('https:'==location.protocol?'//ssl':'//www')+'.google-analytics.com/ga.js';
+            s.parentNode.insertBefore(g,s)}(document,'script'));
+        </script>
+        
+    </body>
+</html>
+
+