+++ section/head-responsive.php
@@ -28,14 +28,15 @@ When you look into an abyss, the abyss also looks into you.
 		<link rel="apple-touch-icon-precomposed" href="img/apple-touch-icon.png"> <!-- Matt -->
 
 		<!-- Mobile View -->     
-		<meta name="viewport" content="width=1024"><!-- Responsive Off -->  
-		<!-- <meta name="viewport" content="width=device-width"> Responsive On -->    
+		<!-- <meta name="viewport" content="width=1025"> Responsive Off -->  
+		<meta name="viewport" content="width=device-width"> <!-- Responsive On -->   
 		
 		<!-- Old Browser Support -->
 		<script src="js/modernizr-2.6.2-respond-1.1.0.min.js"></script>
 		
 		<!-- Style -->
 		<link rel="stylesheet" href="css/base.min.css">
+		<link rel="stylesheet" href="css/responsive.min.css">
         <link rel="stylesheet" href="css/main.css">
         <link rel="stylesheet" href="css/font-awesome.min.css">